+++ test/instream/encoder/influxql_test.exs
@@ -1,5 +1,0 @@
-defmodule Instream.Encoder.InfluxQLTest do
-  use ExUnit.Case, async: true
-
-  doctest Instream.Encoder.InfluxQL, import: true
-end