+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     - Support for implementing `use Instream.Writer` has been removed in favor of `@behaviour Instream.Writer`
     - Support for singular time units has been removed
     - The default JSON decoder has been switched from `:poison` to `:jason`
+    - The query builder has been removed
 
 ## v0.20.0 (2019-04-19)